Move process to its own package
