+++ runner/app/live/streamer/__init__.py
@@ -1,4 +1,3 @@
 from .streamer import PipelineStreamer
-from .process_guardian import ProcessGuardian
 
-__all__ = ["PipelineStreamer", "ProcessGuardian"]
+__all__ = ["PipelineStreamer"]
